Apply ConnectAsync/DisconnectAsync fixes to DI-registered adapter; remove unused Plc.Adapters duplicate
Agent-Logs-Url: https://github.com/efargas/S7-Tools/sessions/6b3076ee-61ea-4021-90c9-d79753c8857c
